requirements.txt
# Core LangGraph and AI dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.50.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=5.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
python-multipart>=0.0.9

# Financial Calculations
quantlib>=1.35.0
scipy>=1.12.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.9

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and Monitoring
structlog>=24.1.0
sentry-sdk[fastapi]>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Development Tools
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# File Processing
python-magic>=0.4.27
chardet>=5.2.0

# Date and Time
python-dateutil>=2.8.2
pytz>=2024.1

# JSON and XML Processing
lxml>=5.0.0
xmltodict>=0.13.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# SWIFT Message Processing
# swift-mt>=0.1.0  # Removed - package not available

# Machine Learning
scikit-learn>=1.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Async Support
asyncio-mqtt>=0.16.0

# Caching
aioredis>=2.0.0

# Background Tasks
celery>=5.3.0

# API Documentation
# fastapi-openapi>=0.1.0  # Removed - package not available
